requirements.txt
Django==5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
uWSGI==2.0.28
pytest-django==4.9.0